Do a stupid amount of work just to throttle multiple FSW changes
This was inspired by https://github.com/benbhall/FileSystemWatcherMemoryCache, but here we use the standalone Microsoft.Extensions.Caching.Memory, as this isn't an ASP.NET Core app.

The API is very different, and this implementation of IMemoryCache *doesn't* automatically fire any event when an item expires (https://github.com/aspnet/Caching/issues/248)... so we have to do some rather nasty stuff with cancellation tokens in order to get a notification on expiry.

We will now only fire a single event, 1 second after the last FSW event of any type (change, rename etc). Hopefully this will keep the amount of indexing activity to a minimum.